RDKEMW-5238: Remove LISA patch and configure persistent path (#690)
* RDKEMW-5238: Remove LISA patch and configure persistent path

Reason for change: Realtek VA requires the installation path for DAC Apps to be /tmp/data so adding variable that can be configured from product layer

Test Procedure: Build and verify.

Risks: Low

Signed-off-by: Jitha James <jitha_james@comcast.com>

* RDKEMW-5238: Update LISA revision and point to no branch

Reason for change: SRCREV to be updated to point to the latest revision. Updated recipe to point CMF_GITHUB_SRC_URI_SUFFIX so as to point to no branch

Test Procedure: Build and verify.

Risks: Low

Signed-off-by: Jitha James <jitha_james@comcast.com>

---------

Signed-off-by: Jitha James <jitha_james@comcast.com>
Co-authored-by: joseinweb <joseinweb@gmail.com>
Co-authored-by: Anand Kandasamy <37086488+anand-ky@users.noreply.github.com>